+++ pkg/webui/console/containers/applications-table/index.js
@@ -255,6 +255,7 @@ const ApplicationsTable = props => {
   return (
     <FetchTable
       entity="applications"
+      defaultOrder="application_id"
       headers={headers}
       addMessage={sharedMessages.addApplication}
       tableTitle={<Message content={sharedMessages.applications} />}